Renames some files to better match our style guide
