+++ lute/runtime/CMakeLists.txt
@@ -4,13 +4,13 @@ target_sources(Lute.Runtime PRIVATE
     include/lute/ref.h
     include/lute/runtime.h
     include/lute/userdatas.h
-    include/lute/UVRequest.h
-    include/lute/UVStream.h
+    include/lute/uvrequest.h
+    include/lute/uvstream.h
     include/lute/uvutils.h
 
     src/ref.cpp
     src/runtime.cpp
-    src/UVRequest.cpp
+    src/uvrequest.cpp
     src/uvutils.cpp
 )
